adding more tests ( old rdvs)
