+++ tests/dommies/rdvs.js
@@ -1,0 +1,52 @@
+const rdvs = [
+	{
+		comment: '',
+		date: 'Thu, 21 Oct 2021 10:00:00 GMT',
+		duration: '15min',
+		id: '172',
+		modality: 'En agence',
+		title: 'Gabriel Adgeg ',
+	},
+	{
+		comment: 'test fix',
+		date: 'Thu, 21 Oct 2021 11:00:00 GMT',
+		duration: '45min',
+		id: '162',
+		modality: 'En agence',
+		title: 'Gabriel Adgeg ',
+	},
+	{
+		comment: 'Hey',
+		date: 'Mon, 25 Oct 2021 12:00:00 GMT',
+		duration: '30min',
+		id: '166',
+		modality: 'Par téléphone',
+		title: 'RémiDormoy',
+	},
+	{
+		comment: 'Test id',
+		date: 'Wed, 27 Oct 2021 12:00:00 GMT',
+		duration: '1h',
+		id: '173',
+		modality: 'Par téléphone',
+		title: 'JuGa',
+	},
+	{
+		comment: 'Test test',
+		date: 'Sun, 31 Oct 2021 13:00:00 GMT',
+		duration: '1h',
+		id: '174',
+		modality: 'En agence',
+		title: 'RémiDormoy',
+	},
+	{
+		comment: '',
+		date: 'Mon, 16 May 2022 09:00:00 GMT',
+		duration: '30min',
+		id: '175',
+		modality: 'Par téléphone',
+		title: 'ElisabethJohnson',
+	},
+]
+
+export { rdvs }
